refactor(billing): address PR feedback
- Merge billing content into member-settings; drop the separate billing
  tab (member-settings was otherwise sparse and billing fits the
  member-profile surface)
- Rename add-card-modal → add-credit-card-modal and update component /
  i18n key references to disambiguate Stripe cards from deck Cards
- Extract Stripe type shapes to a colocated `src/api/billing/stripe.d.ts`
- Extract the Stripe Elements Appearance config to
  `src/utils/billing/stripe-theme.ts`, shared by checkout and
  add-credit-card
- Add JSDoc to every billing query and mutation hook
