+++ src/api/billing/db/index.ts
@@ -1,5 +1,11 @@
 import { supabase } from '@/supabase-client'
 import logger from '@/utils/logger'
+import type {
+  StripeInvoice,
+  StripePaymentMethod,
+  StripeSubscription,
+  StripeUpcomingInvoice
+} from '../stripe'
 
 export type CreateSubscriptionArgs = { planId: string }
 export type CreateSubscriptionResult = {
@@ -22,60 +28,6 @@ export async function createSubscription(
 
   return data
 }
-
-// Minimal Stripe shapes — full SDK types would require shipping @stripe/stripe
-// to the frontend. We read only the fields the UI cares about.
-
-export type StripePrice = {
-  id: string
-  unit_amount: number | null
-  currency: string
-  recurring: { interval: 'day' | 'week' | 'month' | 'year' } | null
-  product: string | { id: string; name: string }
-}
-
-export type StripeSubscriptionItem = {
-  id: string
-  price: StripePrice
-}
-
-export type StripePaymentMethod = {
-  id: string
-  card: {
-    brand: string
-    last4: string
-    exp_month: number
-    exp_year: number
-  } | null
-}
-
-export type StripeSubscription = {
-  id: string
-  status: string
-  current_period_end: number
-  cancel_at_period_end: boolean
-  canceled_at: number | null
-  items: { data: StripeSubscriptionItem[] }
-  default_payment_method: StripePaymentMethod | string | null
-}
-
-export type StripeInvoice = {
-  id: string
-  number: string | null
-  amount_paid: number
-  amount_due: number
-  currency: string
-  status: string | null
-  created: number
-  hosted_invoice_url: string | null
-  invoice_pdf: string | null
-}
-
-export type StripeUpcomingInvoice = {
-  amount_due: number
-  currency: string
-  period_end: number
-} | null
 
 type ManagePayload =
   | { action: 'get-subscription' }
