+++ src/locales/en-us.json
@@ -98,54 +98,53 @@
   "settings.app-settings.accessibility.label": "Accessibility",
   "settings.member-settings.title": "Member Settings",
   "settings.member-settings.general.label": "General",
-  "settings.billing-settings.title": "Billing",
-  "settings.billing-settings.loading": "Loading billing details…",
-  "settings.billing-settings.error": "Couldn't load billing details. Please try again.",
-  "settings.billing-settings.free.label": "Subscription",
-  "settings.billing-settings.free.status": "Free plan",
-  "settings.billing-settings.free.upgrade": "Upgrade",
-  "settings.billing-settings.plan.label": "Plan",
-  "settings.billing-settings.plan.current": "Current plan",
-  "settings.billing-settings.plan.unknown": "Unknown plan",
-  "settings.billing-settings.plan.price": "{amount} {currency}",
-  "settings.billing-settings.plan.price-per-interval": "{amount} {currency} / {interval}",
-  "settings.billing-settings.plan.renews-on": "Renews {date}",
-  "settings.billing-settings.plan.cancels-on": "Cancels {date}",
-  "settings.billing-settings.plan.status.active": "Active",
-  "settings.billing-settings.plan.status.trialing": "Trial",
-  "settings.billing-settings.plan.status.past_due": "Past due",
-  "settings.billing-settings.plan.status.canceled": "Canceled",
-  "settings.billing-settings.plan.status.unpaid": "Unpaid",
-  "settings.billing-settings.plan.status.incomplete": "Incomplete",
-  "settings.billing-settings.plan.status.incomplete_expired": "Expired",
-  "settings.billing-settings.plan.cancel": "Cancel subscription",
-  "settings.billing-settings.plan.cancel-confirm": "Your plan will stay active until the end of the current period.",
-  "settings.billing-settings.plan.cancel-confirm-button": "Confirm cancel",
-  "settings.billing-settings.plan.cancel-abort": "Keep plan",
-  "settings.billing-settings.plan.cancel-success": "Subscription will cancel at period end.",
-  "settings.billing-settings.plan.cancel-error": "Couldn't cancel subscription. Please try again.",
-  "settings.billing-settings.plan.resume": "Resume subscription",
-  "settings.billing-settings.plan.resume-success": "Subscription resumed.",
-  "settings.billing-settings.plan.resume-error": "Couldn't resume subscription. Please try again.",
-  "settings.billing-settings.payment-methods.label": "Payment methods",
-  "settings.billing-settings.payment-methods.empty": "No saved cards.",
-  "settings.billing-settings.payment-methods.expires": "Expires {expiry}",
-  "settings.billing-settings.payment-methods.default": "Default",
-  "settings.billing-settings.payment-methods.make-default": "Make default",
-  "settings.billing-settings.payment-methods.detach": "Remove",
-  "settings.billing-settings.payment-methods.add": "Add card",
-  "settings.billing-settings.payment-methods.set-default-success": "Default card updated.",
-  "settings.billing-settings.payment-methods.set-default-error": "Couldn't update default card.",
-  "settings.billing-settings.payment-methods.detach-success": "Card removed.",
-  "settings.billing-settings.payment-methods.detach-error": "Couldn't remove card.",
-  "settings.billing-settings.invoices.label": "Invoices",
-  "settings.billing-settings.invoices.empty": "No invoices yet.",
-  "settings.billing-settings.invoices.view": "View",
-  "settings.billing-settings.add-card.title": "Add a card",
-  "settings.billing-settings.add-card.loading": "Loading…",
-  "settings.billing-settings.add-card.error": "Couldn't load the card form. Please try again.",
-  "settings.billing-settings.add-card.submit": "Save card",
-  "settings.billing-settings.add-card.submit-error": "Couldn't save card. Please try again.",
+  "settings.member-settings.billing.loading": "Loading billing details…",
+  "settings.member-settings.billing.error": "Couldn't load billing details. Please try again.",
+  "settings.member-settings.billing.free.label": "Subscription",
+  "settings.member-settings.billing.free.status": "Free plan",
+  "settings.member-settings.billing.free.upgrade": "Upgrade",
+  "settings.member-settings.billing.plan.label": "Plan",
+  "settings.member-settings.billing.plan.current": "Current plan",
+  "settings.member-settings.billing.plan.unknown": "Unknown plan",
+  "settings.member-settings.billing.plan.price": "{amount} {currency}",
+  "settings.member-settings.billing.plan.price-per-interval": "{amount} {currency} / {interval}",
+  "settings.member-settings.billing.plan.renews-on": "Renews {date}",
+  "settings.member-settings.billing.plan.cancels-on": "Cancels {date}",
+  "settings.member-settings.billing.plan.status.active": "Active",
+  "settings.member-settings.billing.plan.status.trialing": "Trial",
+  "settings.member-settings.billing.plan.status.past_due": "Past due",
+  "settings.member-settings.billing.plan.status.canceled": "Canceled",
+  "settings.member-settings.billing.plan.status.unpaid": "Unpaid",
+  "settings.member-settings.billing.plan.status.incomplete": "Incomplete",
+  "settings.member-settings.billing.plan.status.incomplete_expired": "Expired",
+  "settings.member-settings.billing.plan.cancel": "Cancel subscription",
+  "settings.member-settings.billing.plan.cancel-confirm": "Your plan will stay active until the end of the current period.",
+  "settings.member-settings.billing.plan.cancel-confirm-button": "Confirm cancel",
+  "settings.member-settings.billing.plan.cancel-abort": "Keep plan",
+  "settings.member-settings.billing.plan.cancel-success": "Subscription will cancel at period end.",
+  "settings.member-settings.billing.plan.cancel-error": "Couldn't cancel subscription. Please try again.",
+  "settings.member-settings.billing.plan.resume": "Resume subscription",
+  "settings.member-settings.billing.plan.resume-success": "Subscription resumed.",
+  "settings.member-settings.billing.plan.resume-error": "Couldn't resume subscription. Please try again.",
+  "settings.member-settings.billing.payment-methods.label": "Payment methods",
+  "settings.member-settings.billing.payment-methods.empty": "No saved cards.",
+  "settings.member-settings.billing.payment-methods.expires": "Expires {expiry}",
+  "settings.member-settings.billing.payment-methods.default": "Default",
+  "settings.member-settings.billing.payment-methods.make-default": "Make default",
+  "settings.member-settings.billing.payment-methods.detach": "Remove",
+  "settings.member-settings.billing.payment-methods.add": "Add card",
+  "settings.member-settings.billing.payment-methods.set-default-success": "Default card updated.",
+  "settings.member-settings.billing.payment-methods.set-default-error": "Couldn't update default card.",
+  "settings.member-settings.billing.payment-methods.detach-success": "Card removed.",
+  "settings.member-settings.billing.payment-methods.detach-error": "Couldn't remove card.",
+  "settings.member-settings.billing.invoices.label": "Invoices",
+  "settings.member-settings.billing.invoices.empty": "No invoices yet.",
+  "settings.member-settings.billing.invoices.view": "View",
+  "settings.member-settings.billing.add-credit-card.title": "Add a card",
+  "settings.member-settings.billing.add-credit-card.loading": "Loading…",
+  "settings.member-settings.billing.add-credit-card.error": "Couldn't load the card form. Please try again.",
+  "settings.member-settings.billing.add-credit-card.submit": "Save card",
+  "settings.member-settings.billing.add-credit-card.submit-error": "Couldn't save card. Please try again.",
   "settings.notification-settings.title": "Notification Settings",
 
   "alert.generic-title": "Continue?",
