+++ tests/integration/phone/apps/settings/component/member-settings.test.js
@@ -1,12 +1,69 @@
-import { describe, test, expect, vi } from 'vite-plus/test'
+import { describe, test, expect, vi, beforeEach } from 'vite-plus/test'
 import { shallowMount } from '@vue/test-utils'
 import { defineComponent, h, useAttrs } from 'vue'
+
+const { mockModalOpen, subscriptionState } = vi.hoisted(() => ({
+  mockModalOpen: vi.fn(),
+  subscriptionState: {
+    plan: 'paid',
+    isLoading: false,
+    error: null,
+    data: { subscription: { id: 'sub_1' }, upcoming: null }
+  }
+}))
 
 vi.mock('@/stores/member', () => ({
   useMemberStore: () => ({
     display_name: 'Alice',
-    email: 'alice@example.com'
+    email: 'alice@example.com',
+    get plan() {
+      return subscriptionState.plan
+    }
   })
+}))
+
+vi.mock('@/api/billing', () => {
+  const noopMutation = () => ({ mutateAsync: vi.fn(), isLoading: { value: false } })
+  const noopQuery = () => ({
+    isLoading: { value: false },
+    error: { value: null },
+    data: { value: null }
+  })
+  return {
+    useSubscriptionQuery: () => ({
+      isLoading: {
+        get value() {
+          return subscriptionState.isLoading
+        }
+      },
+      error: {
+        get value() {
+          return subscriptionState.error
+        }
+      },
+      data: {
+        get value() {
+          return subscriptionState.data
+        }
+      }
+    }),
+    useInvoicesQuery: noopQuery,
+    usePaymentMethodsQuery: noopQuery,
+    useChangePlanMutation: noopMutation,
+    useCancelSubscriptionMutation: noopMutation,
+    useResumeSubscriptionMutation: noopMutation,
+    useSetDefaultPaymentMethodMutation: noopMutation,
+    useDetachPaymentMethodMutation: noopMutation,
+    useCreateSetupIntentMutation: noopMutation
+  }
+})
+
+vi.mock('@/composables/modal', () => ({
+  useModal: () => ({ open: mockModalOpen })
+}))
+
+vi.mock('@/components/modals/checkout.vue', () => ({
+  default: { name: 'Checkout' }
 }))
 
 const SectionHeaderStub = defineComponent({
@@ -26,6 +83,16 @@ const UiInputStub = defineComponent({
   }
 })
 
+const UiButtonStub = defineComponent({
+  name: 'UiButton',
+  inheritAttrs: false,
+  emits: ['click'],
+  setup(_props, { slots, emit }) {
+    const attrs = useAttrs()
+    return () => h('button', { ...attrs, onClick: () => emit('click') }, slots.default?.())
+  }
+})
+
 async function makeMemberSettings() {
   const MemberSettings = (await import('@/phone/apps/settings/component/member-settings.vue'))
     .default
@@ -34,13 +101,34 @@ async function makeMemberSettings() {
     global: {
       stubs: {
         SectionHeader: SectionHeaderStub,
-        UiInput: UiInputStub
+        UiInput: UiInputStub,
+        UiButton: UiButtonStub,
+        PlanSection: {
+          name: 'PlanSection',
+          render: () => h('div', { 'data-testid': 'plan-section-stub' })
+        },
+        PaymentMethodsSection: {
+          name: 'PaymentMethodsSection',
+          render: () => h('div', { 'data-testid': 'payment-methods-stub' })
+        },
+        InvoicesSection: {
+          name: 'InvoicesSection',
+          render: () => h('div', { 'data-testid': 'invoices-stub' })
+        }
       }
     }
   })
 }
 
-describe('member-settings', () => {
+beforeEach(() => {
+  mockModalOpen.mockReset()
+  subscriptionState.plan = 'paid'
+  subscriptionState.isLoading = false
+  subscriptionState.error = null
+  subscriptionState.data = { subscription: { id: 'sub_1' }, upcoming: null }
+})
+
+describe('member-settings — general', () => {
   test('renders the general section', async () => {
     const wrapper = await makeMemberSettings()
     expect(wrapper.find('[data-testid="member-settings__general"]').exists()).toBe(true)
@@ -54,11 +142,53 @@ describe('member-settings', () => {
     expect(values).toContain('Alice')
     expect(values).toContain('alice@example.com')
   })
+})
 
-  test('no longer renders a subscription section (moved to billing tab)', async () => {
+describe('member-settings — billing (free plan)', () => {
+  beforeEach(() => {
+    subscriptionState.plan = 'free'
+  })
+
+  test('renders the free-plan upgrade prompt', async () => {
     const wrapper = await makeMemberSettings()
-    expect(wrapper.find('[data-testid="member-settings__subscription"]').exists()).toBe(false)
-    expect(wrapper.find('[data-testid="member-settings__subscription-free"]').exists()).toBe(false)
-    expect(wrapper.find('[data-testid="member-settings__subscription-paid"]').exists()).toBe(false)
+    expect(wrapper.find('[data-testid="member-settings__free"]').exists()).toBe(true)
+    expect(wrapper.find('[data-testid="plan-section-stub"]').exists()).toBe(false)
+  })
+
+  test('clicking upgrade opens the Checkout modal as a mobile sheet', async () => {
+    const wrapper = await makeMemberSettings()
+    await wrapper.find('[data-testid="member-settings__upgrade"]').trigger('click')
+    expect(mockModalOpen).toHaveBeenCalledWith(expect.objectContaining({ name: 'Checkout' }), {
+      mode: 'mobile-sheet',
+      backdrop: true
+    })
+  })
+})
+
+describe('member-settings — billing (paid plan)', () => {
+  test('renders the loading state while the query resolves', async () => {
+    subscriptionState.isLoading = true
+    const wrapper = await makeMemberSettings()
+    expect(wrapper.find('[data-testid="member-settings__billing-loading"]').exists()).toBe(true)
+    expect(wrapper.find('[data-testid="plan-section-stub"]').exists()).toBe(false)
+  })
+
+  test('renders the error state when the query fails', async () => {
+    subscriptionState.error = new Error('boom')
+    const wrapper = await makeMemberSettings()
+    expect(wrapper.find('[data-testid="member-settings__billing-error"]').exists()).toBe(true)
+    expect(wrapper.find('[data-testid="plan-section-stub"]').exists()).toBe(false)
+  })
+
+  test('renders all three billing sections once loaded', async () => {
+    const wrapper = await makeMemberSettings()
+    expect(wrapper.find('[data-testid="plan-section-stub"]').exists()).toBe(true)
+    expect(wrapper.find('[data-testid="payment-methods-stub"]').exists()).toBe(true)
+    expect(wrapper.find('[data-testid="invoices-stub"]').exists()).toBe(true)
+  })
+
+  test('does not render the free-plan prompt', async () => {
+    const wrapper = await makeMemberSettings()
+    expect(wrapper.find('[data-testid="member-settings__free"]').exists()).toBe(false)
   })
 })
